+++ pkg/roles/dhcp/internal_test_helpers_test.go
@@ -2,14 +2,10 @@ package dhcp
 
 import (
 	"context"
-	"encoding/json"
-	"testing"
 
 	"beryju.io/gravity/pkg/extconfig"
 	"beryju.io/gravity/pkg/roles"
 	"beryju.io/gravity/pkg/storage"
-	"github.com/getsentry/sentry-go"
-	clientv3 "go.etcd.io/etcd/client/v3"
 	"go.uber.org/zap"
 )
 
@@ -59,37 +55,3 @@ func (i *dhcpTestInstance) Migrator() roles.RoleMigrator {
 	return dhcpTestMigrator{}
 }
 
-func setupDHCPInternalTest(t testing.TB) context.Context {
-	t.Helper()
-
-	ctx, cancel := context.WithCancel(t.Context())
-	tx := sentry.StartTransaction(ctx, "test")
-
-	_, err := extconfig.Get().EtcdClient().Delete(tx.Context(), "/", clientv3.WithPrefix())
-	if err != nil {
-		t.Fatalf("failed to reset etcd: %v", err)
-	}
-
-	t.Cleanup(func() {
-		tx.Finish()
-		cancel()
-	})
-
-	return tx.Context()
-}
-
-func mustJSON(v any) string {
-	raw, err := json.Marshal(v)
-	if err != nil {
-		panic(err)
-	}
-	return string(raw)
-}
-
-func panicIfError(args ...any) {
-	for _, arg := range args {
-		if err, ok := arg.(error); ok && err != nil {
-			panic(err)
-		}
-	}
-}
