+++ pkg/tests/apitest/api.go
@@ -1,4 +1,4 @@
-package tests
+package apitest
 
 import (
 	"encoding/base64"
@@ -11,25 +11,26 @@ import (
 	roleAPI "beryju.io/gravity/pkg/roles/api"
 	"beryju.io/gravity/pkg/roles/api/auth"
 	"beryju.io/gravity/pkg/roles/api/types"
+	"beryju.io/gravity/pkg/tests"
 	"github.com/gorilla/securecookie"
 )
 
 func APIClient(rootInst *instance.Instance) (*api.APIClient, func()) {
-	ctx := Context()
+	ctx := tests.Context()
 	inst := rootInst.ForRole("api", ctx)
 	role := roleAPI.New(inst)
-	PanicIfError(role.Start(ctx, []byte(MustJSON(roleAPI.RoleConfig{
+	tests.PanicIfError(role.Start(ctx, []byte(tests.MustJSON(roleAPI.RoleConfig{
 		ListenOverride: "localhost:8008",
 	}))))
 
 	username := base64.RawStdEncoding.EncodeToString(securecookie.GenerateRandomKey(32))
 	token := base64.RawStdEncoding.EncodeToString(securecookie.GenerateRandomKey(64))
 
-	PanicIfError(inst.KV().Put(ctx, inst.KV().Key(
+	tests.PanicIfError(inst.KV().Put(ctx, inst.KV().Key(
 		types.KeyRole,
 		types.KeyUsers,
 		username,
-	).String(), MustJSON(auth.User{
+	).String(), tests.MustJSON(auth.User{
 		Username: username,
 		Permissions: []auth.Permission{
 			{
@@ -38,11 +39,11 @@ func APIClient(rootInst *instance.Instance) (*api.APIClient, func()) {
 			},
 		},
 	})))
-	PanicIfError(inst.KV().Put(ctx, inst.KV().Key(
+	tests.PanicIfError(inst.KV().Put(ctx, inst.KV().Key(
 		types.KeyRole,
 		types.KeyTokens,
 		token,
-	).String(), MustJSON(auth.Token{
+	).String(), tests.MustJSON(auth.Token{
 		Key:      token,
 		Username: username,
 	})))
